only allow explicit preloaded files to work with NtCreateFile
