+++ src/dumpulator/native.py
@@ -11,6 +11,7 @@ from .ntstructs import *
 STATUS_SUCCESS = 0
 STATUS_NOT_IMPLEMENTED = 0xC0000002
 STATUS_INVALID_HANDLE = 0xC0000008
+STATUS_NO_SUCH_FILE = 0xC000000F
 STATUS_ACCESS_DENIED = 0xC0000022
 STATUS_PRIVILEGE_NOT_HELD = 0xC0000061
 STATUS_SET_CONTEXT_DENIED = 0xC000060A  # Return from NtContinue to int 29
